fix: App freezes on backup restoration if the maximum number of clients is reached
fixes https://wearezeta.atlassian.net/browse/SQCORE-498

The previous commit didn't fix the bug in the case described in the title. This one does.
I moved the initial sync to a later phase in the login flow so it doesn't interfere with
removing one of the old clients. Until now, the full sync couldn't finish, because
new client wasn't registered, but the new client couldn't be registered until the backup
restoration screen was up, and the screen was up, because it was waiting for the sync
to finish.