fix: App freeze when the backup is being restored on a new device (#3233)
* fix: App freeze when the backup is being restored on a new device

fixes https://wearezeta.atlassian.net/browse/SQCORE-498

It turns out that asking for a zms instance right after backup restoration but before the end of the initialization process does not work if the app was just installed or if the storage was wiped out. The zms does not exist yet at this point.
(Which is worrying because I remember testing this case a few months ago and it worked back then).

I changed the way of accessing zms to one that shouldn't cause trouble - from the account manager we have just created and we know it exists. I also decided to postpone the sync with backend and to sync only conversations instead of all data. Syncing all data during the backup restoration might actually look like a freeze from the user's point of view, esp. if the internet connection is poor. It's better to do it in the background when the conversation list is loading up. (And then instead of a turning circle the user will see the bar at the top).

* fix: App freezes on backup restoration if the maximum number of clients is reached

fixes https://wearezeta.atlassian.net/browse/SQCORE-498

The previous commit didn't fix the bug in the case described in the title. This one does.
I moved the initial sync to a later phase in the login flow so it doesn't interfere with
removing one of the old clients. Until now, the full sync couldn't finish, because
new client wasn't registered, but the new client couldn't be registered until the backup
restoration screen was up, and the screen was up, because it was waiting for the sync
to finish.