fix: App crash when just after the install the app is put to background (#3242)
fixes https://wearezeta.atlassian.net/browse/SQCORE-600

Since Android 8.1 we need to create a notification channel for our app. We need to do it lazily, but early enough.
The bug was that if the app was opened for the first time, the notification channel was not created right away,
and then when the app was put to background, we called `stopForeground` in `WebSocketService` - that method crashes the app if the notification channel is not created. Fixed that. Now we ensure that the notification channel is
created also when going to the background, before `stopForeground` is called.