feat: Fetch clients for users in batches
Currently we have two ways of fetching clients data:
1. If we send a message to a conversation, the backend may respond that we didn't encrypt it for all clients,
and send us info about missing ones.
2. If the user goes to the single participant info screen, we fetch data about clients of that participant.

With the Federation feature coming in and that we handle bigger and bigger conversation, there is a need for
a more consistent way of fetching client data. The backend gave us a new endpoint /users/list-clients/v2
which we can use for that. Every time we switch to another conversation, we can now ask the backend for all
clients of that conversation in one request. This will speed up sending messages to conversations as there
will be fewer situations when we will receive response that some clients are missing and we have to encrypt
the message again, for them.

Both current ways of fetching clients data are preserved, they will be just used not so often. But they are
still necessary in case if the clients list changed while we are in the conversation.