refactor: Fetch clients for users in batches
This is a change to OtrClient and classes connected to it that will make it possible to ask the backend
for clients of many users at once. Currently we do it for each client separately which brings performance
issues in large conversations. The new implementation is supposed, in future, to use the new endpoint
/list-clients to fetch a dictionary of users and their clients.

However, to make it work, I first need to implement qualifiable ids from the federation feature.
So, even though this PR could be merged as it is, it's really unfinished work. I want to make it into a PR
to just have it as a point of reference, implement qualifiable ids, and then come back to it.

Please do not merge yet.