refactor: Identifiers as newtypes
I'm starting to work on QualifiableIds for the Federation feature - a user id will be replaced in many places by a pair of user id and a domain.
In preparation I looked into our `Uid.scala` file with all ids we use and I decided to refactor it a bit for performance.

Every time we have a `case class SomethingId(str: String)` and we create an instance of it, we create a wrapper over the inner string.
We can avoid that by switching to `final case class SomethingId(str: String) extends AnyVal`. This gives the Scala compiler a hint that it should
strip the wrapper class during the compilation and just leave the string - we save memory and time required for allocation.

Unfortunately, it's not possible everywhere. In case of `ConvId` and `RConvId` - as far as I understand - the compiler first removes the wrappers
and only then try to compile two methods from `CallingServiceImpl`: `updateCallIfActive(rConvId: RConvId)` and `updateCallIfActive(convId: ConvId)`
which leads to an error that they both have the same signature. In case of `UserId`, this class is used in old Java code and that also leads to
a compilation error. But it looks like in all other cases we can have a small performance boost for free.

I didn't change `ContactId` because we don't use it anymore. I will create another PR removing it.