feat: Qualified Ids
Changes to UserInfo and UserData which ensure that Wire Android will use qualified ids where they are available,
but also will work without them. A qualified id is a pair of user id and domain (a string). The backend should
return it as a part of each UserInfo, but it will also return id as a separate field in the json.  In case when
qualified_id exists, we give it priority, but anyway both should be the same.

Additionally, I changed ClientId to a newtype, in accordance to changes of other ids from a recent PR.