fix: A long click on the conversation list doesn't open the bottom dialog (#3297)
The operation of displaying the bottom dialog should be executed on the Ui thread, but was on the Background thread.
This was introduced with recent changes to wire-signals. I looked through the rest of the changes and modified some
of it to make sure that the operations are made on Ui. Especially in case of scrolling.