fix: Glide exception in user notifications (#3292)
fixes https://wearezeta.atlassian.net/browse/SQCORE-51

The issue was described in here https://github.com/wireapp/wire-android/issues/777#issuecomment-531071779
Since then it stopped being so important because we fixed the problem with missing avatars (the notifications were not displayed if the sender had no avatar). So, this fix is basically just a safeguard against it happening again.