fix: Syncing clients on fresh install is broken
The bug is in the code which should update the clients data fetched from the backend. It looks like it was introduced when I changed it
to take advantage from the fact we can do it not one by one client, but in batches, it stopped working at least in the corner case
where we try to send a message immediately after a fresh install.