fix: Syncing clients on fresh install is broken (#3301)
The bug is in the code which should update the clients data fetched from the backend. It looks like it was introduced when I changed it to take advantage from the fact we can do it not one by one client, but in batches, it stopped working at least in the corner case where we try to send a message immediately after a fresh install.

At least in part the bug was also connected with that the endpoint /list-clients/v2 requires qualifiable ids (userId + domain) but it's possible that we won't have the information about the domain for some (or all) users. /list-clients/v2 doesn't accept an empty domain field as valid. In such case we have to revert to the old endpoint as well.

So now, the code in OtrClientsSync tries to do both:

    It splits the set qualifiable ids it gets int two subsets - one for qualifiable ids with domains, and the other for those without.
    It uses the new endpoint (/list-clients/v2) to get information about clients for users with domains.
    It uses the old endpoint (/users/$user/clients) to get information about clients for users without domains.
    If the new endpoint doesn't work it uses the old endpoint as fallback.
    After getting all the information about clients in any way, it updates the clients one by one. This part could be improved - we are capable of updating clients faster - but at least we know it's not bugged.

By the way, @johnxnguyen noticed that we no longer get any information about clients other than clientId and the "class" - a field we don't recognize. This needs to be fixed in another ticket.