fix: Update clients in batches (#3302)
The yesterday fix for sending a first message in a conversation made it harder (or maybe even impossible) to handle large conversations. This is the attempt to merge both these problems. The clients are updated in batches (so, a fix to large conversations) and they are properly synced in new conversations.