fix: Download assets on Android 7 & 9
This PR contains changes to how we download assets to the Android device.
The previous fix didn't work for all devices. It looks like in some cases we are unable to create subfolders from the app
(in our case, we tried to create a "Wire Images" subfolder), but saving an image to the original folder works fine.
When we weren't able to create a subfolder, the method was throwing an exception and aborting the whole operation.
Now it just returns `false`. I checked the other places where it is used - there should be no problems.

I also refactored the code to have the same functionality for saving images and other files with the new and the old API.
This gives me a bit more confidence in the situation when this functionality might work differently depending on the device.