fix: Scroll to position
fixes https://wearezeta.atlassian.net/browse/SQCORE-648

There was a race condtion, probably caused by changes in wire-signals - the code that previously run on the UI thread (and was synchronous
because of that) and looked for the searched message AND then scrolled the conversation to that message, is now split between the Background
thread (the first part) and the UI thread (the second part). I think that scrolling happened before the message was found. That would also
explain why it worked if the user searched for their own message - it was in the memory cache. For other messages, the app had to access
DB and it was a bit slower.

I simplified the code. The old version performed some additional operations which I think have no more use (such as checking the id of
the previous conversation (?)).