fix: Faster retrieval and update of conversation roles
linked to https://wearezeta.atlassian.net/browse/SQCORE-532

After first logging to an account with a lot of conversations a lot of time is used to synchronize them.
In logs I found that a big part of conversation syncing is about retrieving conversation roles and updating them
in the database. I decided to make a few quick optimizations:
1. We don't need to retrieve conversation roles for non-team conversations - they are always as default.
2. When updating, we need to find the conversation by remote id. It was done in O(n). Now we first make a hashmap
with remote ids as keys, and look for conversations in it, so most of the time it's O(1).

I also changed the logic a bit to use default roles when we are unable to retrieve them. Until now, this was
treated as an error.