Federation: Add domain to the search query
https://wearezeta.atlassian.net/browse/SQCORE-672

SearchQuery class gets a new field - the domain. If the user searches for other users as before, using only the name or the handle,
nothing changes. But now it's possible as well to add the domain to the search query string using the format: "<username>@<domain>"
or "@<handle>@<domain>". This means that we want to look for the user <username> (or @<handle>) only on the given domain.