Federation: Proper handling of responses to user search queries
Adding a domain to search queries in a previous PR already affected the user search functionality
and introduced a few bugs. This PR updates the user search functionality with handling of queries
and search results with domains. Some changes were also made in preparation of the next step -
adding connections with a user from another domain.