fix: Empty query in user search returns zero results
This is a very simple bug but it was obscured by another bug on the backend which also affected search results.
Before, `UserData.matchQuery(query: SearchQuery): Boolean` returned `true` if the query was empty, but that was
misleading - it should have returned true only if the user matched the query after all. I changed that but forgot
that our user search actually depends on it, the rule being that with the empty query we should return all local
results.
Now, I simply  moved the check for empty query to the user search logic.