fix: Bump down wire-signals to 0.4.3 to fix logging in
wire-signals 0.5.1 contains a change in how signals are initialized. I thought that changes was minor and in fact improves
the consistency of signals behaviour. Now I'm not so sure anymore. I will have to take another look in the wire-signals code
and meanwhile bumping down to 0.4.3 will be a workaround.