wire-signals version bump to 0.5.2
In signals we have an optimization that if there is no subscriber to a signal (i.e. no `.foreach`, `.head`, `.onUi`, etc.
- calls to methods that create subscriptions) then the signal might decide that the intermediate computations, done in
methods like `.map`, are not necessary, because there is nobody who would consume their results. (That's why we should be
carefult with putting side-effects in `.map`, but that's not the issue here).

There was an inconsistency in SourceSignal - if created with an `apply` method, this optimization in the new source signal
was disabled. But if created with `new SourceSignal`, it was on, In wire-signals 0.5.1 I decided to remove the inconsistency
by making all ways to create a source signal disable the optimization. It looked harmless enough and I thought it's safer
this way. However, to the contrary, it turns out our logging in flow actually relies on that the source signal will not
compute its state immediately, but only after some time later in the flow. If the computation is done too early, the value
of the signal `account` in `MainActivity` will be `None` and the flow will decide to move the user back to the logging in
screen. Only if the value is not computed (because there are no subscriptions) until the moment we need it, the computation
will result in the valid account manager and the user will progress in the flow.

This points at a problem with the logging in flow. We shouldn't rely on when the computation occurs. But for now I'm not able
to look into it - the flow is quite complex - so I decided that instead I will remove the inconsistency the other way around.
In wire-signals 0.5.2, every way we create a source signal, the optimization is on. I looked through the code, and tested
the app, and I think it's safe, but automatic tests are needed to ensure that the UI (where we have a lot of source signals
waiting for input from the user) works fine.