fix: Search results are empty when User searches on group details page (#3382)
fixes https://wearezeta.atlassian.net/browse/SQCORE-727

The main cause of the bug was that UserData.matchesQuery tried to check if the domain of the matched user is the same
as the domain of the current user and that doesn't work without the rest of the Federation feature in place - we would
need to be able to get the domain of the other user from the backend. For now, I removed the check for domain, so
matchesQuery works like before.

I also add a "throttle" on the filter signal. Currently the user search among conversation participants is performed
with each keystroke. For large conversations that may be too CPU-heavy, especially that users usually type a few
characters before expecting the search to work. A throttle, set to 500 milliseconds, means that if the user will
quickly type a few characters, the search will be performed only when it's finished.

And I used this opportunity to add the "final" keyword to classes in the ParticipantAdapter file, as well as to
correct indentation, etc.