fix: Partner inside of team can't be found (#3386)
fixes https://wearezeta.atlassian.net/browse/SQCORE-786

Similar to a previous bug about user search (https://github.com/wireapp/wire-android/pull/3382) this one too
was because the exact match search tried to compare domains. Without the ability to learn about the domain
of the remote user, which is not implemented yet, this always failed, resulting in that the external/partner
user was always filtered out from the user search.