Federation: Open a conversation with a federated user
A partial implementation of https://wearezeta.atlassian.net/browse/SQCORE-674
Adding a user to an existing conversation will be implemented in the next commit.

I deleted an old unused file, PickUserDataState.java . I thought about reimplementing the old PickUserController
from Java to Scala, as that would help me a bit, but in the end it looks like it would take too much time.