Add users from fake 1-to-1 conversations to the results of user search
This is needed for now so that we will be able to search for federated users in the local data.
Currently, every conversation with a federated user is a fake 1-to-1 group conversations and they are
treated as if they were unconnected with the current user. This should change in the future
so maybe this code could be deleted, but it also works in general as a way to include fake 1-to-1
conversations in our user search, so maybe it could stay.