fix: conversation-join deep link doesn't work
fixes https://wearezeta.atlassian.net/browse/SQSERVICES-638

The problem was that on Android we expected the deep link with the format
wire://conversation-join/?key=<key>&code=<code>
And on other platforms it's
wire://conversation-join?key=<key>&code=<code>
without the slash / after conversation-join

We needed that / to properly parse the deep link. I made some quick changes to make it work.
It would be nice to write it better in the future.