fix: The user can create a conversation with another user that blocked them (#3416)
fixes https://wearezeta.atlassian.net/browse/SQCORE-646

It's also a simplification of how group and fake 1-to-1 conversations are created. From now on, a group conversation
will be created only with the author as the initial participant, and all others will be added separately.
This way it's much easier to control what happens if the conversation creation failed or when adding a user failed.