Federation: Add the conversation domain and the sender ("from") domain to events
On new backends each conversation event will contain data about the domain its coming from:
the domain of the conversation and also the domain of the sender, if applicable.
This PR is mostly just rewriting the Event class where after adding these two fields a lot of other code
had to be changed. I wrote JSON parsing in such way that if for some reason the Android app is connected
to the old backend which sends events without domains, the event will be handled properly and
the domains will be set to None.

Changes in other classes are only side-effects, like in pattern matching where now we need to take into
account that there are more fields in the given event class.