fix: Create a fake 1-to-1 conversation (#3438)
fixes https://wearezeta.atlassian.net/browse/SQCORE-848

After recent changes in creating group conversations we now first create an empty conversation and then add members to it. When we add new members, we check if the current user has the right to do it. But now we have a corner case: we have an empty conversation where the self user is not yet an active user and, since it's empty, the conversation is not yet recognized as a group conversation. So the check returns false, preventing us from adding members.
This PR introduces a flag "forceModifyMembers" to go around the check if we are in this special situation.