fix: Creating conversation fails because of invalid payload
In the last release we changed the way group conversations are created.
The list of participant ids is no longer sent in ConversationInitState. The group conversation starts empty (only with the creator)
and the all other participants are added in a "member join" event. It simplifies handling situations when some users
cannot be added to a conversation for various reasons.
The json for ConversationInitState was also changed, because of work on Federation - if the list of users is empty (as it always is now for groups)
then we skip the field. This works for our testing environment, so we didn't catch it as a bug, but on the backend used
by one of our customers this fails with ErrorResponse(code: 400, message: Error in $: key "users" not found, label: invalid-payload).

The hotfix reverts the change for non-federated setups. If the federation feature flag is off, we will still add the "users" field to the json,
even if it's empty.