fix: revert to old way of adding users to conversations if the federation flag is off
The customer is still not able to create a new conversation. Another possibility is that we try to add members to
the conversation in a way that works only with new endpoints. This commit adds a check for the federation flag
in a few places and if the federation is off, it reverts to old methods.

The amount of code changed is mostly due to indentation - the previous code is now wrapped in if/else-s.