fix: Use old endpoints when the federation feature flag is off (#3488)
This commit introduces more careful separation of the federation feature.
In the last release we changed the way group conversations are created. The list of participant ids is no longer sent in ConversationInitState. The group conversation starts empty (only with the creator) and the all other participants are added in a "member join" event. It simplifies handling situations when some users cannot be added to a conversation for various reasons. The json for ConversationInitState was also changed, because of work on Federation - if the list of users is empty (as it always is now for groups) then we skip the field. On the backend used by one of our customers this fails with `ErrorResponse(code: 400, message: Error in $: key "users" not found, label: invalid-payload)`.

This PR reverts the change: If the federation feature flag is off, we will still add the "users" field to the json, even if it's empty. Also, it reverts to using the old code and endpoints in a few cases.

The amount of code changed is mostly due to indentation - the previous code is now wrapped in if/else-s.

This is a version of https://github.com/wireapp/wire-android/pull/3485 and https://github.com/wireapp/wire-android/pull/3486 - those PRs are a hotfix on top of the release branch. This PR here introduces the same changes to the federation feature.