Federation: Disable promoting people from other backends to administrators
For now it's not possible to promote another participant to the group conversation admin role.
The switch on the single participants view is disabled (it's still visible though).
Also, I fixed a small bug in the single participants view - the left bottom button was "Connect"
for federated users, because for now they're not really connected to us, they're more like
team members.
Both these solutions are temporary.