Federation: Disable promoting people from other backends to administrators (#3498)
* Federation: Disable promoting people from other backends to administrators

For now it's not possible to promote another participant to the group conversation admin role. The switch on the single participants view is disabled (it's still visible though). Also, I fixed a small bug in the single participants view - the left bottom button was "Connect" for federated users, because for now they're not really connected to us, they're more like team members. Both these solutions are temporary.

* Federation: Remove a federated user from the group conversation (#3500)

Uses a new endpoint for removal of a user on federated backends and for self-removal from a group conversation.
During testing I experienced a few "glitches" when join/remove events were not processed until a subsequent event came (e.g. a message). I tried to debug it but it stopped happening after I added a few log statements. I decided to keep the new log statement should the issue happen again.