Federation: Use new endpoits for retrieving conversations data (#3494)
* Federation: Use new endpoits for retrieving conversations data

Together with Federation we change the way we communicate with the backend. For now only for federated backends but in the future this will be extended to all, as far as I understand. Right now there are two ways of getting conversation data from the backend:
1. If we have a convId, we can call `GET /conversations/ids=...`
2. If we want all conversations, we can simply call `GET /conversations` without ids
The second version will return a list of conversations data + info if there is more and we should call again with
a special parameter indicating where we stopped.

This will change now. Instead we will separate endpoinds, one for getting conversation ids in this paginating manner, just a the second version above, but in response we will get only ids (`POST /conversations/list-ids`). And then we can use those ids to request conversation data from another endpoint (`POST /conversations/list/v2`). We can use the second endpoint directly if we already know the ids.

* fixing unit tests and deleting missing convs from the local storage
* simplify conversation json deserializers