Federation: Display the domain in the handle
If the user data contains the domain, the handle (or "username") of that user should look like this: `@personalhandle@domain`
instead of just `@personalhandle`. But editing the handle is still possible only for the personal handle.
I also added "Domain" and "Team" fields to the Account view. They're not editable. (the "Team" does not belong to the Federation
feature, but it was missing, while the design docs show it just next to "Domain").