fix: Creating a new conversation fails (#3509)
At some point of working on federation I made a mistake in the name of the class which should be used in deserialization of the response (a json object) returned from the backend when we create a new conversation. The faulty code not only didn't fail at compilation but also at runtime the app just ignored the error and failed at a different place, giving a wrong error message. It took me quite a while to figure out what's actually wrong.

What's even worse, this code is written in a way that makes it really hard to unit test it. I can't think of any way
right now that wouldn't be a few days of work of preparing mocks and actually testing the tests. It was supposed
to be type-safe which would make unit tests unnecessary. As we know now, it's not.