fix: User sees enterprise alert when starting 1:1 call in a non paying team
fixes https://wearezeta.atlassian.net/browse/SQSERVICES-842

The alert should be visible only for real group conversations, that is, if there are more than two participants
or if the name of the conversation is customized, indicating that even though there are only two people in,
it's still considered a group conv, not a fake 1-to-1.
The bug was because we checked explicitely if the conversation type is GROUP, but this is not enough.