fix: When switching conversations, the contents of the previous conversaton is visible for a while
The process of refreshing contents of the current conversation and the process of opening the conversation view are asynchronous.
It means that the situation described in the title can happen - it's just that usually it should be so short that almost
unnoticeable. It looks like the Federation feature added additional computations to the refereshing contents and so now it's
easier to see this glitch (even if the feature flag is off).

I made a few changes to solve it.
1. I removed one operation from the process - checking which message was read last - and I added it later when the contents
   are already displayed. When this operation was included, after opening the conversation, the last read operation changed
   automatically, and then the whole refreshing was repeated because of that. (Note that the scrolling mechanism seems to broken anyway, so
   checking the last read message in order to scroll to that message doesn't work).
2. I added a check to opening the conversation to make it wait until the messages of the new conversation are loaded.
   It doesn't solve the problem completely, but it makes the glitch considerably shorter.
3. Additionally, I delayed opening the convo my 100ms, which gives the content refreshing process a slightly higher priority.

I know it all sounds a bit hacky, but I can't spend more time on this problem, so this is the best I could do.