Federation: 1-to-1 connections (#3524)
Connecting to a user from another backend should work just as connecting to someone on the same backend.
Right now the makeshift solution is to immediately create a fake 1-to-1 conversation between users from different backends. The new approach will be to go through the whole connection process, with sending a request and waiting for confirmation.