Federation: Don't show the domain in the user handle if it's the same as the current domain
Until now the domain was always displayed.
The large number of modified files is a result of that the new requirement forces us to check the current domain,
which is saved in the user data of the current user - we need to access that with a `Future` which turns the whole
operation Future-based.