fix: Unresponsive message view after fresh start
This is a partial revert of "fix: When switching conversations, the contents of the previous conversaton is visible for a while (#3533)".
It turns out that commit resulted in another bug where the message view becomes unresponsive to tapping and the animations are broken
after a fresh start of the app.

This reverts commit c5e071ca46e619b62bca9c88c7edc99d4ce17c13.