Federation: Methods for qualified conversations in ConversationsClient
A bit of duplicated code that works on federated versions of endpoints.
For now those methods are not used - I will plug them in slowly, one by one, in the given places.
This is just my attempt to cut that work into smaller pieces.
I also made a few changes to the old code in ConversationsClient to make names more consistent across the class.