fix: When searching for users to add to a conversation, display only connected or team users
The bug allowed for trying to add to a conversation a user that was not connected to us.
The adding failed on the backend side, and on Android we got an error because of that,
but the unconnected user shouldn't be dsplayed in the search results in the first place.