fix: Unconnected users recognized as federated on SignleParticipantViews
For some time, as a temporary solution, we allowed to open a conversation with a federated user as if we were already connected.
That meant that on the SingleParticipantView we displayed the "Open conversation" button instead of "Connect" when the other user
was not connected, but from a federated backend. This was recently removed as we now have proper connections functionality
in Federation. But that removal caused a bug: now every unconnected user was recognized as federated and so the "open conversation"
button was displayed.

This PR removes the check for "isFederated" (which was a temporary solution and should have been already removed before) and
also makes sure that in other situation when the "isFederated" check is being made, it will always return false if the Federation
feature flag is off.