hotfix: Federation: Session Reset
I forgot about one use case - session reset between devices on different backends. I implemented sessions with qualified ids
while it was still not possible to reset them, and later tested them only on the same backend (the reset works in the old
way even if the user is on a federated backend but the same as the other user).