Federation: Display the "federated users are present" in the conversation banner
Just as we already have banners in conversations that say "guests are present" and "services are active",
we want to display information about federated users in the conversation, and about special users - "externals".
Also, all the combinations of those four categories should be displayed accordingly.

"Federated users" is the new category but "externals" were also missing from the banner so I added them now.

The commit contains also some refactoring.