Notifications: Decrypt messages
https://wearezeta.atlassian.net/browse/SQCORE-1139

After fetching messages from the backend and storing them in the db, we decrypt them in-place, that is,
we replace the encrypted blob in the db with its decrypted counterpart, and set the flag 'decrypted' to true.
This partially solves SQCORE-1140 as well.

This PR also disables PushService in the app from trying to handle the messages parallel to what we are doing here.
In SQCORE-1142 this functionality will be turned on again, but this time it will take over only when the app
comes to the front.

Plus some refactoring and unit tests.