fix: Conference calling user preference and the enterprise alert
https://wearezeta.atlassian.net/browse/SQSERVICES-906

A change in how we handle the conference calling user preference.
This should fix both that on on-premise installation with older backends conference calling should stil be enabled
and that the user for whom conference calling is disabled should see the enterprise alert only when it becomes
enabled by a change on the backend.

Also this commit contains fixes to unit tests, broken recently by some other changes.