fix: Conference calling user preference and the enterprise alert SQSERVICES-906 (#3600)
* fix: Conference Calling flag works only for new users

Setting the flag to `Some(true)` should also work for users who updated. For them, the default value
is `None`. We need a change in the property's name to reset that.

* fix: Conference calling user preference and the enterprise alert

https://wearezeta.atlassian.net/browse/SQSERVICES-906

A change in how we handle the conference calling user preference.
This should fix both that on on-premise installation with older backends conference calling should stil be enabled
and that the user for whom conference calling is disabled should see the enterprise alert only when it becomes
enabled by a change on the backend.

Also this commit contains fixes to unit tests, broken recently by some other changes.

* update pipeline to reenable zmessage tests

* a fix to the situation when the backend returns an error

* making all feature flags handle the backend error the same way

Co-authored-by: Marco <gardosen@gmx.de>