Notifications: A missed call notification (SQCORE-1180)
https://wearezeta.atlassian.net/browse/SQCORE-1180

This is a corner case. When the user misses a call, we should display a notification about it.
But all other notifications come to us from the backend - messages, pings, likes, user connection
requests, etc. This one is local. We should create it as a response to a callback from AVS saying
that the call was missed. But because there is no event, and only that callback, we need another
way of making the notifications system aware that such a notification should be created.