feat: Notifications: Show notifications when websockets are always on (SQCORE-1192)
https://wearezeta.atlassian.net/browse/SQCORE-1192

The user may decide not to use FCM and keep websockets always on, even when the app is in the background.
In such case the notification should appear on the screen as well.

In the old implementation this was done by NotificationService in the same way as when a new event comes
by FCM. Now notifications are handled separately and the new implementation does not work when the websockets
are on. To solve this, I reintroduced NotificationService and plugged it back to the event handling process
just as before, but now it's a very simple class, called WsNotificationService to show the difference.
It is supposed to work only when the websockets are on *and* the ap is in the background. It parses the event
with NotificationParser and then passes it forward to UI.

I also fixed a problem with PushService that might have resulted in that some incoming events were ignored.
Currently when we receive a new event through websockets, we store it quickly in the database, and then
we retrieve it to decrypt it, but the retrieval is from the cache, not from the db directly. It looks like
in some cases the cache might have not yet updated and we didn't get the event.