feat: Notifications: Fix notifications sounds (SQCORE-1182)
https://wearezeta.atlassian.net/browse/SQCORE-1182

The functionality was simplified. Now we will have only four notification channels: incoming calls, outgoing calls, messages, and pings.
There will be no differentiation based on user accounts - it gives little value and as the current implementation shows, it introduces
a lot of complexity and we were never able to handle it well.