fix: Ensure notifications are being cancelled (SQCORE-1199)
https://wearezeta.atlassian.net/browse/SQCORE-1199

When a notification appears on the screen, the user should be able to cancel (or “dismiss”) them. There are two ways to do it:
1. The user may swipe the notification right.
2. The user may open a conversation (either through clicking on the notification or by simply going to the conversation) and then
   the notifications connected to that conversation should be cancelled.

On top of that, if a new notification appears while the old one is still visible (or if it was cancelled before) the old notification
should not appear again, as if it was a new one.