feat: Notifications: Show notifications when the app is open but on a different conversation (SQCORE-1197)
This is to restore a feature of the old implementation.
Even when the app is open, and so the events come through websockets, we still want to display a notification
in the case when the user has the app open on one conversation, and the new message comes to another conversation.