feat: Make sure the new system works with Federation
https://wearezeta.atlassian.net/browse/SQCORE-1181

The only issue connected to Federation that I found in the new notifications is about user connection events.
A user connection from another federated server requires to fetch user data and for that we need to use the qualified id
of the other person.
I update the code to use the qualified id if Federation is on.

By the way, I think right now there is a problem in general with user connections on Federation, not connected to new notifications.
I will look into it in another PR.