fix: Notifications: More aggressive cancelling of notifications (SQCORE-1199)
https://wearezeta.atlassian.net/browse/SQCORE-1199

Too often notifications stay in the notifications tray or reappear there even after they are swiped away by the user.
The only sure way to get rid of them is to go to the conversation they are from. But when you get a lot of notifications
from many conversations, this could be tedious.

In this PR I fix the way the app notices when we open it and when we switch from the conversation list to a conversation and back.
Thanks to this it's now possible to remove all notifications for the given user account (not only for one conversation) when
the user goes to the conversation list.
Also, I added a line where all the notifications from the user account are cancelled when the user logs out - that's important
for QA tests.